Refactored FAQ retrival pipeline to reduce code duplication
